requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.0.3
datasets==2.0.0
dill==0.3.5.1
docker-pycreds==0.4.0
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.48.1
huggingface-hub==0.9.1
idna==3.4
importlib-metadata==4.12.0
install==1.3.5
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.13
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
promise==2.3
protobuf==3.19.5
psutil==5.9.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-lightning==1.6.0
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.12.1
torch==1.11.0+cu113
torchmetrics==0.9.3
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
Werkzeug==2.2.2
xxhash==3.0.0
yacs==0.1.8
yarl==1.8.1
zipp==3.8.1
